Initialize Turbolinks so initial state is 'remembered'
Turbolinks doesn't remember the initial url.

Steps to reproduce.
1. Visit initial page in new tab to ensure new history stack.

    `http://localhost:9292/test/index.html`

2. Click on __Other page__ link.
3. Click on _Home_ link.
4. Hit back once. You should be at `other.html`
5. Hit back again. The url changes to `index.html` but the page stays the same.

Hit forward and back a couple of times to see that the url changes but the page stays the same.

The reason this happens is because line 95 executes only when the `turbolinks` state data is present which currently isn't for the initial page.