Add ability to mark <head> tags with data-turbolinks-replace.
This will remove the tag on page change and replace it with the new version from the freshly fetched page.

This is useful for example with a csrf-token stored in a session that might expire. Or could be used instead of data-turbolinks-track to replace CSS instead of doing a full page reload.